+++ packages/core/src/eager.ts
@@ -7,7 +7,5 @@ export const eager = <Value>(
   lazyPromise: LazyPromise<Value, unknown>,
 ): Promise<Value> =>
   new Promise((resolve, reject) =>
-    lazyPromise.subscribe(resolve, reject, () => {
-      reject(new Error(`The LazyPromise passed to eager(...) has failed.`));
-    }),
+    lazyPromise.subscribe(resolve, reject, reject),
   );